requirements.txt
fastapi
uvicorn
openai
langchain
langchain-huggingface
pdfplumber
python-docx
extract-msg
faiss-cpu
sentence-transformers
python-dotenv
